BLO-351 move navbar + related components to their own folder
